--- packages/extension/e2e/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@argent-x/e2e",
-  "version": "5.2.0",
-  "private": true,
-  "repository": "github:argentlabs/argent-x",
-  "homepage": "https://github.com/argentlabs/argent-x/#readme",
-  "devDependencies": {
-    "@playwright/test": "^1.29.0"
-  },
-  "scripts": {
-    "format": "prettier --loglevel warn --write \"**/*.{js,jsx,ts,tsx,css,md,yml,json}\"",
-    "lint": "eslint . --cache --ext .ts",
-    "test:e2e": "playwright test",
-    "prepare": "yarn tsc"
-  },
-  "license": "GPL-3.0-only"
-}
