+++ packages/extension/playwright.config.ts
@@ -1,28 +1,48 @@
+import path from "path"
+
 // playwright.config.ts
 import type { PlaywrightTestConfig } from "@playwright/test"
-import { devices } from "@playwright/test"
 
-const config: PlaywrightTestConfig = {
+import config from "./e2e/src/config"
+
+const isCI = Boolean(process.env.CI)
+
+const playwrightConfig: PlaywrightTestConfig = {
+  workers: 1,
   timeout: 5 * 60e3, // 5 minutes
   reportSlowTests: {
     threshold: 2 * 60e3, // 2 minutes
     max: 5,
   },
-  forbidOnly: !!process.env.CI,
-  testDir: "./e2e",
-  retries: process.env.CI ? 2 : 0,
+  expect: { timeout: 60 * 1000 }, // 1 minute
+  reporter: isCI
+    ? [
+        ["github"],
+        [
+          "json",
+          { outputFile: path.join(config.reportsDir, "extension.json") },
+        ],
+        ["list"],
+        [
+          "html",
+          {
+            open: "never",
+            outputFolder: path.join(config.reportsDir, "playwright-report"),
+          },
+        ],
+      ]
+    : "list",
+  forbidOnly: isCI,
+  testDir: "e2e/src/specs",
+  testMatch: /\.spec.ts$/,
+  retries: isCI ? 2 : 0,
   use: {
     trace: "on-first-retry",
+    viewport: { width: 360, height: 600 },
+    actionTimeout: 60 * 1000, // 1 minute
   },
-  projects: [
-    {
-      name: "Plugin view",
-      use: {
-        ...devices["Desktop Chrome"],
-        viewport: { width: 360, height: 600 },
-      },
-    },
-  ],
+  outputDir: config.artifactsDir,
+  preserveOutput: isCI ? "failures-only" : "never",
 }
 
-export default config
+export default playwrightConfig
